Moved functionality for adding featured instrumentation to its own module
